Change crate type to lib
This makes it possible to use shvcall from external crates in tests.
